+++ NEWS.md
@@ -1,3 +1,11 @@
+# ICD10gm 1.2.5
+This is a minor update:
+
+- Data for the year 2023
+- Three new interactive lookup functions (icd_lookup, icd_browse, icd_search)
+- Replaced outdated URLs in the "Coding the Pandemic" vignette with archive.org snapshots
+
+
 # ICD10gm 1.2.4
 This is a minor update: